requirements.txt
mysql-connector-python
PyYAML